* disable implicit conversion for all TimeRef operators. Fixes a number of bugs. * add a semicolon after RELAYTOOL_XXX to not confuse kdevelops c++ parser * Start to extract resample code from ResampleAudioReader into a dedicated class Resampler (not functional yet) * Move Mixer.h/cpp to common dir

+++ src/commands/SplitClip.cpp
@@ -49,7 +49,7 @@ SplitClip::SplitClip(AudioClipView* view)
 
 int SplitClip::prepare_actions()
 {
-	if (m_splitPoint == 0) {
+	if (m_splitPoint == qint64(0)) {
 		m_splitPoint = TimeRef(cpointer().scene_x() * m_sv->timeref_scalefactor);
 	}
